pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "bili_stalker_mcp"
version = "3.0.0"
description = "MCP Server for Bilibili data retrieval with lightweight lists and detail tools"
authors = [
    {name = "222wcnm", email = "2328072813li@gmail.com"},
]
readme = {file = "README.md", content-type = "text/markdown"}
requires-python = ">=3.12"
license = "MIT"
keywords = ["mcp", "bilibili", "stalker", "video", "updates", "ai", "model-context-protocol"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Multimedia :: Video",
    "Topic :: Communications",
]
dependencies = [
    "bilibili-api-python>=17.1.4",
    "fastmcp>=3.0.0",
    "mcp>=1.26.0",
    "httpx>=0.25.0",
    "async-lru>=2.0.4",
    "pydantic>=2.7.0",
    "curl_cffi>=0.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.0.0",
    "python-dotenv>=1.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[project.scripts]
bili-stalker-mcp = "bili_stalker_mcp.cli:main"

[project.urls]
Homepage = "https://github.com/222wcnm/BiliStalkerMCP"
Documentation = "https://github.com/222wcnm/BiliStalkerMCP#readme"
Repository = "https://github.com/222wcnm/BiliStalkerMCP"
Issues = "https://github.com/222wcnm/BiliStalkerMCP/issues"
Changelog = "https://github.com/222wcnm/BiliStalkerMCP/releases"

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.0.0",
    "python-dotenv>=1.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["bili_stalker_mcp"]

[tool.black]
line-length = 88
target-version = ['py312']
include = '''\.pyi?'''

extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["bili_stalker_mcp"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

